refactor Resource namespace
